ai-service/requirements.txt
flask
groq
python_dotenv
